Further Sonar fixes and API change (#50)
* Further sonar fixes and adding CMake custom target for analyze
* Further sonar resolutions and coverages. Changed api from /api/v1/ready to /api/v1/ping with response including appmanager state
* Adding fixes/sonar to github actions branch list
* Preparing for sonar analyze on GH actions

+++ src/Config.hpp
@@ -37,7 +37,7 @@ namespace nucleus::config
     extern std::string pool_main_file;
 
     // For ReSTServer
-    extern int rest_port;
+    extern unsigned short rest_port;
     extern std::string rest_address;
     extern size_t rest_threads;
